requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
psycopg2-binary==2.9.9
Pillow==10.1.0
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
